// src/emno/types/types.ts
export interface EmnoConfig {
  baseUrl?: string;
  token: string;
}

export interface ErrorBody {
  message: string;
  detail?: string;
}

export const AlgoEnum = z.enum(["l2", "ip", "cosine"]);

export const ErrorResponseSchema = z.object({
  message: z.string(),
  detail: z.any().optional(),
});

export const CollectionConfigSchema = z.object({
  dim: z.number(),
  m: z.number().optional(),
  efConstruction: z.number().optional(),
  ef: z.number().optional(),
  model: z.string().optional().nullable(),
  algo: z.string().optional().nullable(),
  allowReplace: z.boolean().optional().nullable(),
});
export type CollectionConfigSchemaType = z.infer<typeof CollectionConfigSchema>;

export const CreateCollectionRequestSchema = z.object({
  name: z.string(),
  description: z.string().optional().nullable(),
  config: CollectionConfigSchema,
});

export const CreateCollectionResponseSchema = z.object({
  name: z.string(),
  description: z.string().optional().nullable(),
  config: CollectionConfigSchema,
  id: z.string().optional().nullable(),
});

export const DeletedCollectionResponseSchema = z.object({
  deleted: CreateCollectionResponseSchema,
});

export const GetAllCollectionResponseSchema = z.array(
  CreateCollectionResponseSchema,
);

export const UpdateCollectionRequestSchema = z.object({
  name: z
    .string()
    .regex(/^[a-zA-Z0-9_-]+$/)
    .optional(),
  description: z.string().optional(),
});

export const VectorSchema = z.array(z.number());

export const VectorObjectSchema = z.object({
  values: VectorSchema,
  metadata: z.any().optional(),
  content: z.string().nonempty(),
});

export const VectorObjectWithIdSchema = z.object({
  values: VectorSchema,
  metadata: z.any().optional(),
  content: z.string().nonempty(),
  id: z.string().nullable(),
});

export const VectorObjectForUpsertSchema = z.object({
  values: VectorSchema.optional(),
  metadata: z.any().optional(),
  content: z.string().optional(),
  id: z.string().nullable().optional(),
});

export const VectorCreateListRequestSchema = z.array(
  z.object({
    values: VectorSchema,
    metadata: z.any().optional(),
    content: z.string().nonempty(),
  }),
);

export const VectorListResponseSchema = z.array(
  z.object({
    value: z.array(z.number()).optional().nullable(),
    metadata: z.any().optional(),
    content: z.string().nonempty(),
    id: z.string().nullable(),
  }),
);

export const VectorListUpsertResponseSchema = z.object({
  created: z.array(
    z.object({
      values: z.array(z.number()).optional().nullable(),
      metadata: z.any().optional(),
      content: z.string().nonempty().optional(),
      id: z.string().nullable(),
    }),
  ),
  updated: z.array(
    z.object({
      values: z.array(z.number()).optional().nullable(),
      metadata: z.any().optional(),
      content: z.string().nonempty().optional(),
      id: z.string().nullable(),
    }),
  ),
  errored: z.array(z.any()),
});

export const DeleteVectorListRequestSchema = z.object({
  ids: z.array(z.string()).optional(),
  deleteAll: z.boolean().optional(),
});

export const DeleteVectorListResponseSchema = z.object({
  deleted: VectorListResponseSchema,
});

export const GetVectorListRequestSchema = z.object({
  ids: z.array(z.string()).min(1),
  includeVectorValues: z.boolean().optional(),
});

export const VectorUpdateListRequestSchema = z.array(
  VectorObjectForUpsertSchema,
);
// ...
export const QueryVectorListSchema = z.object({
  vectors: z.array(VectorSchema),
  limit: z.number().optional(),
});

export const QueryVectorListResponseSchema = z.array(
  z.array(
    z.object({
      id: z.string().nullable(),
      content: z.string(),
      metadata: z.any(),
      distance: z.number(),
    }),
  ),
);
// ...
export const CountVectorsResponseSchema = z.object({ count: z.number() });